fsm-dp: LLC support
LLC support for the fsm-dp driver. A new MHI channel is defined for
LLC transport. There are two MHI transport channels per fsm-dp driver
instance.  Internally, the ring structure is extended to support a
second set of ring buffers, and ring pointers for LLC channels.
Messages received from a MHI transport channel are put on the
corresponding ring to separate and prioritize traffic. For DL traffic,
an ioctl is extended to the application to specify which transport
channel the message should be sent on.

Change-Id: If7424502a282aef315ebaef126d62e834941c226
CRs-Fixed: 3114383
Signed-off-by: Che-Min Hsieh <quic_cheminh@quicinc.com>