fsm_dp: UL and DL traffic profiling
Add UL and DL traffic profiling. Each UL or DL packet is time-stamped at
certain life cycle of the packet. At end of life cycle of a DL packet, which
is DMA transfer complete of a packet, the saved timestamps of the
the packet are collected to a circular buffer.
A debugfs node is added to retrieve and analyze the saved circular buffer.

For the UL packet, packets are time-stamped at DMA completion, and sending
of receive indication time. However, the rest of the timestamping and the
collection of timestamps are done in the applcation fsm-dp library.

Change-Id: I96f9fd381207a80e4b0552b358ada25e80820ebc
CRs-Fixed: 3264159
Signed-off-by: Che-Min Hsieh <quic_cheminh@quicinc.com>